Create Armstrong Numbers
Find out if a given number is an "Armstrong Number".

An n-digit number that is the sum of the nth powers of its digits is called an n-Armstrong number. Examples :
371 = 3**3 + 7**3 + 1**3;
9474 = 9**4 + 4**4 + 7**4 + 4**4;
93084 = 9**5 + 3**5 + 0**5 + 8**5 + 4**5.

Write a Python program that;
takes a positive integer number from the user,
checks the entered number if it is Armstrong,
consider the negative, float and any entries other than numeric values then display a warning message to the user.