Create Ceasar Cipher
Our employer asked us for a program that encrypts a user-entered text. The required conditions are:

In the entered Turkish texts, each letter will be converted to the next two letters in alphabetical order. (For example, 'a' will be 'c', 'z' will be 'b').
If there is a non-letter character in the entered text, 'Please enter a text consisting of letters only!' After the warning, input will be requested again.
Uppercase letters will be converted to uppercase and lowercase letters to lowercase.
Whitespaces will be kept.