Create Palindromic Sentence
Write a function/functions that checks whether the sentence you get from the user is a palindrome. (Do not consider punctuation and special characters. Only consider "alphanumeric" characters.)

input : "Was it a car or a cat I saw?"

output : "Was it a car or a cat I saw?" is a palindrome