Create Rangoli Alphabet
You are given a letter. Your task is to print an alphabet rangoli of letter index. (Rangoli is a form of Indian folk art based on creation of patterns.)

Different sizes(letter indexes) of alphabet rangoli are shown below:

#size 2

----c----
--c-b-c--
c-b-a-b-c
--c-b-c--
----c----

#size 4

--------e--------
------e-d-e------
----e-d-c-d-e----
--e-d-c-b-c-d-e--
e-d-c-b-a-b-c-d-e
--e-d-c-b-c-d-e--
----e-d-c-d-e----
------e-d-e------
--------e--------

#size 9

------------------j------------------
----------------j-i-j----------------
--------------j-i-h-i-j--------------
------------j-i-h-g-h-i-j------------
----------j-i-h-g-f-g-h-i-j----------
--------j-i-h-g-f-e-f-g-h-i-j--------
------j-i-h-g-f-e-d-e-f-g-h-i-j------
----j-i-h-g-f-e-d-c-d-e-f-g-h-i-j----
--j-i-h-g-f-e-d-c-b-c-d-e-f-g-h-i-j--
j-i-h-g-f-e-d-c-b-a-b-c-d-e-f-g-h-i-j
--j-i-h-g-f-e-d-c-b-c-d-e-f-g-h-i-j--
----j-i-h-g-f-e-d-c-d-e-f-g-h-i-j----
------j-i-h-g-f-e-d-e-f-g-h-i-j------
--------j-i-h-g-f-e-f-g-h-i-j--------
----------j-i-h-g-f-g-h-i-j----------
------------j-i-h-g-h-i-j------------
--------------j-i-h-i-j--------------
----------------j-i-j----------------
------------------j------------------
The center of the rangoli has the first alphabet letter a, and the boundary has the letter index(in alphabetical order).

Function Description

Complete the rangoli function in the editor below.

rangoli has the following parameters:

str letter: the letter to return its rangoli regard to alphabetical index
Returns

string: a single string made up of each of the lines of the rangoli separated by a newline character (\n)
Credentials: HackerRank